requirements.txt
selenium
pandas
requests
python-dotenv
conda-forge::googletrans
deep-translator
chromium-browser
chromium-chromedriver